requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
langchain
langchain-community
langchain_openai
langgraph.checkpoint.sqlite
